Address pr review
